Split `mc_asm` to different modules.
