+++ mc_asm/src/to_asm.rs
@@ -1,187 +1,17 @@
 use ordered_float::OrderedFloat;
-use std::collections::BTreeMap;
-use std::collections::HashMap;
-use std::collections::HashSet;
 use std::collections::VecDeque;
-use std::fmt;
 use std::mem;
 
 use mc_ir::{Arg, IntermediateRepresentation, Op};
 use mc_parser::ast::*;
 
-#[derive(Debug)]
-pub struct Asm {
-  lines: Vec<String>,
-  labels: BTreeMap<usize, String>,
-  strings: BTreeMap<String, String>,
-  floats: BTreeMap<OrderedFloat<f64>, String>,
-  builtin_functions: HashSet<Identifier>,
-}
-
-impl fmt::Display for Asm {
-  fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-    for line in &self.lines {
-      line.fmt(f)?;
-      writeln!(f)?;
-    }
-
-    Ok(())
-  }
-}
+use crate::asm::Asm;
+use crate::stack::*;
+use crate::storage::*;
+use crate::register::*;
 
 pub trait ToAsm {
   fn to_asm(&self) -> Asm;
-}
-
-#[derive(Debug)]
-pub struct Stack {
-  temporary_register: BTreeMap<usize, Reg32>,
-  temporaries: VecDeque<Reg32>,
-  lookup_table: HashMap<usize, (usize, bool, bool)>,
-  parameters: Vec<(StorageType, usize, usize)>,
-  parameters_size: usize,
-  stack_size_index: usize,
-  variables: Vec<(StorageType, usize, usize)>,
-  variables_size: usize,
-  used_registers: BTreeMap<Reg32, usize>,
-}
-
-impl Default for Stack {
-  fn default() -> Stack {
-    Stack {
-      temporary_register: Default::default(),
-      temporaries: VecDeque::from(vec![Reg32::EAX, Reg32::EDX, Reg32::ECX, Reg32::EBX, Reg32::EDI, Reg32::ESI]),
-      lookup_table: Default::default(),
-      parameters: Default::default(),
-      parameters_size: 8,
-      stack_size_index: Default::default(),
-      variables: Default::default(),
-      variables_size: Default::default(),
-      used_registers: Default::default(),
-    }
-  }
-}
-
-impl Stack {
-  pub fn lookup(&self, index: usize) -> (StorageType, usize, usize, bool, bool) {
-    let &(i, parameter, array) = self.lookup_table.get(&index).unwrap();
-
-    let (ty, count, offset) = if parameter { self.parameters[i] } else { self.variables[i] };
-
-    (ty, count, offset, parameter, array)
-  }
-
-  pub fn push(
-    &mut self,
-    index: usize,
-    storage_type: StorageType,
-    count: usize,
-    parameter: bool,
-    array: bool,
-  ) -> Pointer {
-    if parameter {
-      let offset = self.parameters_size;
-      self.parameters.push((storage_type, count, self.parameters_size));
-      self.lookup_table.insert(index, (self.parameters.len() - 1, true, array));
-
-      if array {
-        self.parameters_size += StorageType::Dword.size();
-      } else {
-        self.parameters_size += count * storage_type.size();
-      }
-
-      Pointer { base: Reg32::EBP, storage_type: StorageType::Dword, offset, index_offset: None, parameter, array }
-    } else {
-      self.variables_size += count * storage_type.size();
-      self.variables.push((storage_type, count, self.variables_size));
-      self.lookup_table.insert(index, (self.variables.len() - 1, false, array));
-      Pointer { base: Reg32::EBP, storage_type, offset: self.variables_size, index_offset: None, parameter, array }
-    }
-  }
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq, PartialOrd, Ord)]
-pub enum Reg32 {
-  EAX,
-  ECX,
-  EDX,
-  EBX,
-  ESP,
-  EBP,
-  ESI,
-  EDI,
-}
-
-impl fmt::Display for Reg32 {
-  fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-    match self {
-      Self::EAX => "eax",
-      Self::EBX => "ebx",
-      Self::ECX => "ecx",
-      Self::EDX => "edx",
-      Self::ESP => "ebp",
-      Self::EBP => "ebp",
-      Self::EDI => "edi",
-      Self::ESI => "esi",
-    }
-    .fmt(f)
-  }
-}
-
-impl Reg32 {
-  pub fn as_reg8(&self) -> (Reg8, Reg8) {
-    match self {
-      Self::EAX => (Reg8::AL, Reg8::AH),
-      Self::ECX => (Reg8::CL, Reg8::CH),
-      Self::EDX => (Reg8::DL, Reg8::DH),
-      Self::EBX => (Reg8::BL, Reg8::BH),
-      _ => unreachable!(),
-    }
-  }
-}
-
-#[derive(Debug)]
-pub enum Reg8 {
-  AL,
-  CL,
-  DL,
-  BL,
-  AH,
-  CH,
-  DH,
-  BH,
-}
-
-impl fmt::Display for Reg8 {
-  fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-    match self {
-      Self::AL => "al",
-      Self::CL => "cl",
-      Self::DL => "dl",
-      Self::BL => "bl",
-      Self::AH => "ah",
-      Self::CH => "ch",
-      Self::DH => "dh",
-      Self::BH => "bh",
-    }
-    .fmt(f)
-  }
-}
-
-fn push_storage_temporary(storage: Storage, stack: &mut Stack) {
-  if let Storage::Register(_, reg) = storage {
-    push_temporary(reg, stack);
-  }
-}
-
-fn push_temporary(temporary: Reg32, stack: &mut Stack) {
-  stack.used_registers.remove(&temporary);
-
-  if temporary == Reg32::EAX || temporary == Reg32::EDX || temporary == Reg32::ECX || temporary == Reg32::EBX {
-    stack.temporaries.push_front(temporary);
-  } else {
-    stack.temporaries.push_back(temporary);
-  }
 }
 
 macro_rules! stack_hygiene {
@@ -222,140 +52,6 @@ macro_rules! stack_hygiene {
       $stack.temporary_register.insert($i, temp_l);
     }
   };
-}
-
-#[derive(Debug, Clone, Copy, PartialEq)]
-pub enum StorageType {
-  Qword,
-  Dword,
-  Byte,
-}
-
-impl From<&Ty> for StorageType {
-  fn from(ty: &Ty) -> Self {
-    match ty {
-      Ty::Float => Self::Dword,
-      Ty::Int => Self::Dword,
-      Ty::Bool => Self::Byte,
-      Ty::String => Self::Dword,
-    }
-  }
-}
-
-impl fmt::Display for StorageType {
-  fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-    match self {
-      Self::Qword => "QWORD PTR",
-      Self::Dword => "DWORD PTR",
-      Self::Byte => "BYTE PTR",
-    }
-    .fmt(f)
-  }
-}
-
-impl StorageType {
-  fn size(&self) -> usize {
-    match self {
-      Self::Qword => 8,
-      Self::Dword => 4,
-      Self::Byte => 1,
-    }
-  }
-
-  fn map_register(&self, reg: &Reg32) -> String {
-    match self {
-      Self::Dword => reg.to_string(),
-      Self::Byte => reg.as_reg8().0.to_string(),
-      _ => unreachable!(),
-    }
-  }
-}
-
-#[derive(Debug, Clone)]
-pub struct Pointer {
-  base: Reg32,
-  storage_type: StorageType,
-  offset: usize,
-  index_offset: Option<Offset>,
-  parameter: bool,
-  array: bool,
-}
-
-impl fmt::Display for Pointer {
-  fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-    write!(f, "{} [{}", self.storage_type, self.base)?;
-
-    if self.offset != 0 {
-      if self.parameter { "+" } else { "-" }.fmt(f)?;
-
-      write!(f, "{}", self.offset)?;
-    }
-    if let Some(index_offset) = &self.index_offset {
-      write!(f, "+{}*{}", index_offset, self.storage_type.size())?;
-    }
-
-    write!(f, "]")
-  }
-}
-
-#[derive(Debug, Clone)]
-pub enum Storage {
-  Pointer(Pointer),
-  Register(StorageType, Reg32),
-  Fpu(StorageType),
-  Literal(StorageType, String),
-  Label(StorageType, String, bool),
-}
-
-impl fmt::Display for Storage {
-  fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-    match self {
-      Self::Pointer(pointer) => write!(f, "{}", pointer),
-      Self::Register(storage_type, reg) => {
-        if f.alternate() {
-          write!(f, "{}", storage_type.map_register(reg))
-        } else {
-          write!(f, "{}", reg)
-        }
-      }
-      Self::Fpu(_) => unreachable!(),
-      Self::Literal(_, string) => write!(f, "{}", string),
-      Self::Label(ty, label, flat) => {
-        if *flat {
-          write!(f, "OFFSET FLAT:{}", label)
-        } else {
-          write!(f, "{} {}", ty, label)
-        }
-      }
-    }
-  }
-}
-
-#[derive(Debug, Clone)]
-enum Offset {
-  Register(Reg32),
-  Literal(usize),
-}
-
-impl fmt::Display for Offset {
-  fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-    match self {
-      Self::Register(reg) => reg.fmt(f),
-      Self::Literal(lit) => lit.fmt(f),
-    }
-  }
-}
-
-impl Storage {
-  fn storage_type(&self) -> StorageType {
-    match *self {
-      Self::Pointer(Pointer { storage_type, .. }) => storage_type,
-      Self::Register(storage_type, _) => storage_type,
-      Self::Fpu(storage_type) => storage_type,
-      Self::Literal(storage_type, _) => storage_type,
-      Self::Label(storage_type, ..) => storage_type,
-    }
-  }
 }
 
 fn calc_index_offset(stack: &mut Stack, asm: &mut Asm, reg: Reg32, arg: &Arg<'_>) -> Storage {
